+++ examples/simple_browser.py
@@ -4,5 +4,5 @@ import webview
 
 if __name__ == '__main__':
     # Create a standard webview window
-    window = webview.create_window('Simple browser', 'https://webglsamples.org/aquarium/aquarium.html')
+    window = webview.create_window('Simple browser', 'https://pywebview.flowrl.com/hello')
     webview.start()