+++ tests/test_start.py
@@ -17,7 +17,8 @@ def test_start():
 
 
 def assert_func(window):
-    sleep(1)
+    window.events.loaded.wait()
+    print(window.evaluate_js('document.body.innerHTML'))
     html_result = window.evaluate_js('document.getElementById("heading").innerText')
     assert html_result == 'Hello there!'