+++ webview/platforms/edgechromium.py
@@ -317,8 +317,8 @@ class EdgeChrome:
         if request.headers == original_headers:
             return
 
-        missing_headers = {k: v for k, v in request.headers.items() if k not in original_headers or original_headers[k] != v}
-        extra_headers = {k: str(v) for k, v in original_headers.items() if k not in request.headers}
+        extra_headers = {k: v for k, v in request.headers.items() if k not in original_headers or original_headers[k] != v}
+        missing_headers = {k: str(v) for k, v in original_headers.items() if k not in request.headers}
 
         for k, v in extra_headers.items():
             args.Request.Headers.SetHeader(k, v)