+++ webview/window.py
@@ -167,6 +167,7 @@ class Window:
         self.events.loaded = Event(self)
         self.events.before_load = Event(self, True)
         self.events.before_show = Event(self, True)
+        self.events.initialized = Event(self, True)
         self.events.shown = Event(self)
         self.events.minimized = Event(self)
         self.events.maximized = Event(self)
@@ -182,7 +183,7 @@ class Window:
         self.dom = DOM(self)
         self.gui = None
         self.native = None # set in the gui after window creation
-        self.state = State(self)
+        self._state = State(self)
 
     def _initialize(self, gui, server: http.BottleServer | None = None):
         self.gui = gui
@@ -209,6 +210,9 @@ class Window:
         )
         self.real_url = self._resolve_url(self.original_url)
 
+        abort = self.events.initialized.set(gui.renderer)
+        return not abort
+
     @property
     def width(self) -> int:
         self.events.shown.wait(15)
@@ -220,6 +224,16 @@ class Window:
         self.events.shown.wait(15)
         _, height = self.gui.get_size(self.uid)
         return height
+
+    @property
+    def state(self) -> State:
+        return self._state
+
+    @state.setter
+    def state(self, state: State) -> None:
+        if not isinstance(state, State):
+            raise TypeError('State must be an instance of State class')
+        self._state = state
 
     @property
     def title(self) -> str:
@@ -252,13 +266,6 @@ class Window:
         self.__on_top = on_top
         if hasattr(self, 'gui') and self.gui != None:
             self.gui.set_on_top(self.uid, on_top)
-
-    @_loaded_call
-    def get_elements(self, selector: str) -> list[Element]:
-        logger.warning(
-            'This function is deprecated and will be removed in future releases. Use window.dom.get_elements() instead'
-        )
-        return self.dom.get_elements(selector)
 
     @_shown_call
     def load_url(self, url: str) -> None: